@@ -14,7 +14,7 @@
   },
   "devDependencies": {
     "@changesets/cli": "^2.27.1",
-    "eslint": "^8.57.0",
+    "eslint": "^9.20.1",
     "prettier": "^3.2.5",
     "turbo": "^2.4.0"
   },
