@@ -19,7 +19,7 @@
     "@digital-nomad/tsconfig": "workspace:*",
     "@types/react": "^18.2.61",
     "@types/react-dom": "^18.2.19",
-    "eslint": "^8.57.0",
+    "eslint": "^9.20.1",
     "@digital-nomad/eslint-config": "workspace:*",
     "tsup": "^8.0.2",
     "typescript": "5.5.4"
